fix(rpc-provider): move nock to devDependencies so it no longer ships
