+++ packages/rpc-provider/package.json
@@ -31,11 +31,11 @@
     "@polkadot/x-ws": "^14.0.3",
     "eventemitter3": "^5.0.1",
     "mock-socket": "^9.3.1",
-    "nock": "^13.5.5",
     "tslib": "^2.8.1"
   },
   "devDependencies": {
-    "@substrate/connect": "0.8.11"
+    "@substrate/connect": "0.8.11",
+    "nock": "^13.5.5"
   },
   "optionalDependencies": {
     "@substrate/connect": "0.8.11"
