+++ packages/rpc-provider/tsconfig.build.json
@@ -7,6 +7,7 @@
     "resolveJsonModule": true
   },
   "exclude": [
+    "**/*.manual.ts",
     "**/*.spec.ts",
     "**/mod.ts"
   ],
