+++ packages/rpc-provider/tsconfig.spec.json
@@ -9,6 +9,7 @@
     "noEmit": true
   },
   "include": [
+    "**/*.manual.ts",
     "**/*.spec.ts"
   ],
   "references": [
